pldm: Remove unnecessary type casting in logs
This commit removes the unnecessary type casting of values of the
journal logs based on supported types mentioned in LG2 documentation
[1].

Testing: Verified the debug traces to correct.

'''
For Instance:
Earlier -
error(”TYPE = {TYPE} “, “TYPE”, static_cast<unsigned>(PLDM_STATE_EFFECTER_PDR));
Journal trace -
May 27 08:16:51 p10bmc pldmd[931]: TYPE = 11

After Correction -
error(”TYPE = {TYPE}“, “TYPE”, PLDM_STATE_EFFECTER_PDR);
Journal trace -
May 27 08:16:51 p10bmc pldmd[931]: TYPE = Enum(11)
'''

[1]: https://github.com/openbmc/phosphor-logging/blob/master/docs/structured-logging.md#lg2

+++ requester/request.hpp
@@ -90,7 +90,7 @@ class RequestRetryTimer
         if (rc)
         {
             error("Failed to stop the request timer, response code '{RC}'",
-                  "RC", static_cast<int>(rc));
+                  "RC", rc);
         }
     }
 
@@ -197,7 +197,7 @@ class Request final : public RequestRetryTimer
         {
             error(
                 "Failed to send pldmTransport message, response code '{RC}' and error - {ERROR}",
-                "RC", static_cast<int>(rc), "ERROR", errno);
+                "RC", rc, "ERROR", errno);
             return PLDM_ERROR;
         }
         return PLDM_SUCCESS;
