+++ softoff/softoff.cpp
@@ -70,7 +70,7 @@ SoftPowerOff::SoftPowerOff(sdbusplus::bus_t& bus, sd_event* event,
             if (rc != PLDM_SUCCESS)
             {
                 error("Failed to get Sensor PDRs, response code '{RC}'", "RC",
-                      lg2::hex, static_cast<int>(rc));
+                      lg2::hex, rc);
                 hasError = true;
                 return;
             }
@@ -296,7 +296,7 @@ int SoftPowerOff::hostSoftOff(sdeventplus::Event& event)
         instanceIdDb.free(pldmTID, instanceID);
         error(
             "Failed to encode set state effecter states request message, response code '{RC}'",
-            "RC", lg2::hex, static_cast<int>(rc));
+            "RC", lg2::hex, rc);
         return PLDM_ERROR;
     }
 
@@ -337,7 +337,7 @@ int SoftPowerOff::hostSoftOff(sdeventplus::Event& event)
         {
             error(
                 "Failed to receive pldm data during soft-off, response code '{RC}'",
-                "RC", static_cast<int>(rc));
+                "RC", rc);
             return;
         }
 
@@ -363,7 +363,7 @@ int SoftPowerOff::hostSoftOff(sdeventplus::Event& event)
         if (response->payload[0] != PLDM_SUCCESS)
         {
             error("Getting the wrong response, response code '{RC}'", "RC",
-                  (unsigned)response->payload[0]);
+                  response->payload[0]);
             exit(-1);
         }
 
@@ -399,7 +399,7 @@ int SoftPowerOff::hostSoftOff(sdeventplus::Event& event)
         instanceIdDb.free(pldmTID, instanceID);
         error(
             "Failed to send message/receive response, response code '{RC}' and error - {ERROR}",
-            "RC", static_cast<int>(rc), "ERROR", errno);
+            "RC", rc, "ERROR", errno);
         return PLDM_ERROR;
     }
